feat(getos): allow to override the release file
The lsb_release program upstream allows to use the LSB_OS_RELEASE environment variable
to override the release file to consider, it would be nice to support this here.

For instance, we plan to move away from /etc/upstream-release/lsb-release in elementary OS
to /usr/lib/upstream-os-release

Note that the lsb-release file (and its format) is now legacy and isn't even used by
lsb_release itself.